# pkg/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Sandbox0 API
  version: 0.1.0
  description: Public HTTP APIs exposed by internal-gateway.
servers:
  - url: https://api.sandbox0.ai
    description: Production
...
paths:
...
  /api/v1/sandboxes/{id}/contexts:
    post:
      tags: [contexts]
      summary: Create a context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContextRequest"
      responses:
        "201":
          description: Context created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextResponse"
    get:
      tags: [contexts]
      summary: List contexts
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Context list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextListResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}:
    get:
      tags: [contexts]
      summary: Get context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      responses:
        "200":
          description: Context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextResponse"
    delete:
      tags: [contexts]
      summary: Delete context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      responses:
        "200":
          description: Context deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessDeletedResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SandboxID:
      name: id
      in: path
      required: true
      schema:
        type: string
    ContextID:
      name: ctx_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data: {}
      required: [success]
...
    SuccessDeletedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                deleted:
                  type: boolean
...
    ProcessType:
      type: string
      enum: [repl, cmd]
    PTYSize:
      type: object
      properties:
        rows:
          type: integer
          format: int32
        cols:
          type: integer
          format: int32
    ExecCandidate:
      type: object
      properties:
        name:
          type: string
        args:
          type: array
          items:
            type: string
      required: [name]
    REPLEnvVar:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        value_from:
          type: string
      required: [name]
    REPLPromptConfig:
      type: object
      properties:
        custom_prompt:
          type: string
    REPLReadyMode:
      type: string
      enum: [prompt_token, startup_delay]
    REPLReadyConfig:
      type: object
      properties:
        mode:
          $ref: "#/components/schemas/REPLReadyMode"
        token:
          type: string
        startup_delay_ms:
          type: integer
          format: int32
    REPLConfig:
      type: object
      properties:
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/ExecCandidate"
        env:
          type: array
          items:
            $ref: "#/components/schemas/REPLEnvVar"
        default_term:
          type: string
        prompt:
          $ref: "#/components/schemas/REPLPromptConfig"
        ready:
          $ref: "#/components/schemas/REPLReadyConfig"
      required: [name, candidates]
    CreateREPLContextRequest:
      type: object
      properties:
        language:
          type: string
        input:
          type: string
        repl_config:
          $ref: "#/components/schemas/REPLConfig"
    CreateCMDContextRequest:
      type: object
      properties:
        command:
          type: array
          items:
            type: string
        expose_port:
          type: integer
          format: int32
        expose_auto_wakeup:
          type: boolean
    CreateContextRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/ProcessType"
        repl:
          $ref: "#/components/schemas/CreateREPLContextRequest"
        cmd:
          $ref: "#/components/schemas/CreateCMDContextRequest"
        wait_until_done:
          type: boolean
        cwd:
          type: string
        env_vars:
          type: object
          additionalProperties:
            type: string
        pty_size:
          $ref: "#/components/schemas/PTYSize"
        idle_timeout_sec:
          type: integer
          format: int32
        ttl_sec:
          type: integer
          format: int32
    ContextResponse:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ProcessType"
        language:
          type: string
        cwd:
          type: string
        env_vars:
          type: object
          additionalProperties:
            type: string
        running:
          type: boolean
        paused:
          type: boolean
        created_at:
          type: string
        output:
          type: string
        public_url:
          type: string
        exposed_port:
          type: integer
          format: int32
      required: [id, type, running, paused, created_at]
...
    SuccessContextResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ContextResponse"
    SuccessContextListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                contexts:
                  type: array
                  items:
                    $ref: "#/components/schemas/ContextResponse"
...